+++ src/DateTimeFormatter.php
@@ -3,105 +3,65 @@
 namespace Knp\Bundle\TimeBundle;
 
 use Symfony\Contracts\Translation\TranslatorInterface;
-use DateTime;
-use DatetimeInterface;
 
-class DateTimeFormatter
+final class DateTimeFormatter
 {
-    protected $translator;
-
-    public function __construct(TranslatorInterface $translator)
+    /**
+     * @internal
+     */
+    public function __construct(private TranslatorInterface $translator)
     {
-        $this->translator = $translator;
     }
 
     /**
-     * Returns a formatted diff for the given from and to datetimes
+     * Returns a formatted diff for the given from and to datetimes.
      */
-    public function formatDiff(DateTimeInterface $from, DateTimeInterface $to = null, string $locale = null): string
-    {
-        if (null === $to) {
-            $to = new \DateTimeImmutable('now');
-        }
+    public function formatDiff(
+        int|string|\DateTimeInterface $from,
+        int|string|\DateTimeInterface $to = null,
+        string $locale = null
+    ): string {
+        $from = self::formatDateTime($from);
+        $to = self::formatDateTime($to);
 
-        static $units = array(
+        static $units = [
             'y' => 'year',
             'm' => 'month',
             'd' => 'day',
             'h' => 'hour',
             'i' => 'minute',
-            's' => 'second'
-        );
+            's' => 'second',
+        ];
 
         $diff = $to->diff($from);
 
         foreach ($units as $attribute => $unit) {
             $count = $diff->$attribute;
+
             if (0 !== $count) {
-                return $this->doGetDiffMessage($count, $diff->invert, $unit, $locale);
+                $id = sprintf('diff.%s.%s', $diff->invert ? 'ago' : 'in', $unit);
+
+                return $this->translator->trans($id, ['%count%' => $count], 'time', $locale);
             }
         }
 
-        return $this->getEmptyDiffMessage($locale);
+        return $this->translator->trans('diff.empty', [], 'time', $locale);
     }
 
-    /**
-     * Returns the diff message for the specified count and unit
-     *
-     * @param  integer $count  The diff count
-     * @param  boolean $invert Whether to invert the count
-     * @param  string $unit   The unit must be either year, month, day, hour,
-     *                         minute or second
-     */
-    public function getDiffMessage(int $count, bool $invert, string $unit): string
+    private static function formatDateTime(int|string|\DateTimeInterface|null $value): \DateTimeInterface
     {
-        if (0 === $count) {
-            throw new \InvalidArgumentException('The count must not be null.');
+        if ($value instanceof \DateTimeInterface) {
+            return $value;
         }
 
-        $unit = strtolower($unit);
-
-        if (!in_array($unit, array('year', 'month', 'day', 'hour', 'minute', 'second'))) {
-            throw new \InvalidArgumentException(sprintf('The unit \'%s\' is not supported.', $unit));
+        if (is_int($value)) {
+            $value = date('Y-m-d H:i:s', $value);
         }
 
-        return $this->doGetDiffMessage($count, $invert, $unit);
-    }
-
-    /**
-     * Returns a DateTime instance for the given datetime
-     *
-     * @param mixed $dateTime
-     */
-    public function getDateTimeObject($dateTime = null): DateTimeInterface
-    {
-        if ($dateTime instanceof DateTimeInterface) {
-            return $dateTime;
+        if (null === $value) {
+            $value = 'now';
         }
 
-        if (is_int($dateTime)) {
-            $dateTime = date('Y-m-d H:i:s', $dateTime);
-        }
-
-        if (null === $dateTime) {
-            $dateTime = 'now';
-        }
-
-        return new DateTime($dateTime);
-    }
-
-    protected function doGetDiffMessage(int $count, bool $invert, string $unit, string $locale = null): string
-    {
-        $id = sprintf('diff.%s.%s', $invert ? 'ago' : 'in', $unit);
-
-        return $this->translator->trans($id, array('%count%' => $count), 'time', $locale);
-    }
-
-    /**
-     * Returns the message for an empty diff
-     */
-    public function getEmptyDiffMessage(string $locale = null): string
-    {
-        return $this->translator->trans('diff.empty', array(), 'time', $locale);
+        return new \DateTime($value);
     }
 }